\name{update_terms}

\alias{update_terms}

\title{
Update the accepted terms
}

\description{
Update the accepted terms (variables and values) from a repo. 
}


\usage{
update_terms(voc, quiet=FALSE, force = FALSE, local_terms = NULL)
}
 
\arguments{
  \item{voc}{character. path to the github repo of the vocabulary. E.g. "reagro/terminag"}
  \item{quiet}{logical. If \code{TRUE} a message is printed}
  \item{force}{logical}
  \item{local_terms}{character (path)}
} 
 
\value{
none
}

