\name{accepted_variables}

\alias{accepted_variables}

\title{
Get accepted variables
}

\description{
Get a data.frame with accepted variable names
}

\usage{
accepted_variables(voc="carob-data/terminag", include=NULL)
}
 
\arguments{
  \item{voc}{character. path to the github repo of the vocabulary. E.g. "carob-data/terminag"}
  \item{include}{character or NULL. Can be used to use a subset of variables groups}
}

\value{
character
}

